# contracts/api/books/v1/openapi.yaml
openapi: 3.0.3
info:
  title: LoreWeave Books API
  version: 1.5.0
  description: |
    Module 02 — books, chapters (MVP text/plain .txt uploads), editable chapter draft (UTF-8) with revision history,
    cover image, summary, and per-user storage usage. Multilingual: each locale is a separate chapter row with
    original_language; list chapters supports optional query filters original_language and/or sort_order and
    pagination (limit/offset/total). Owner list includes visibility for direct badge rendering in workspace.
    Logical delete: DELETE book/chapter moves to recycle bin (trashed); DELETE …/purge from bin marks purge_pending for
    future garbage collection (blobs + DB rows removed later by a batch job, not in this API). GET /v1/books lists active
    books only; GET /v1/books/trash lists trashed books. Trashed or purge_pending books must not appear in catalog/unlisted/public.
    Raw upload bytes remain in S3-compatible storage (e.g. MinIO); draft/revisions in Postgres until purged. Gateway-facing REST only.
servers:
  - url: http://localhost:3000
    description: API Gateway (BFF)
...
paths:
  /v1/epub-imports/inspect:
    post:
      tags: [imports]
      summary: Upload and inspect an EPUB source without creating chapters
      operationId: inspectEpubImport
      description: >
        Retains the EPUB source for an authenticated user, validates the archive, and returns
        metadata, cover candidates, and an EPUB3-nav/NCX/spine-normalized navigation preview.
        This endpoint does not create a book, chapter, or import job.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: EPUB 2 or EPUB 3 source file, subject to deployment archive limits.
                target_book_id:
                  type: string
                  format: uuid
                  description: Optional existing target. When supplied, edit access is required.
      responses:
        '200':
          description: Valid source inspected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpubInspectionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '415':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          description: EPUB is invalid, encrypted, or violates an archive safety limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ErrorResponse:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    EpubImportWarning:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - navigation_missing_spine_fallback
            - navigation_anchor_missing
            - navigation_range_ambiguous
            - navigation_node_non_content
            - unsupported_resource
            - asset_missing
            - asset_invalid_mime
            - internal_link_target_missing
            - metadata_conflict
            - composition_materialization_failed
            - rollback_conflict_user_modified
        message:
          type: string
        source_key:
          type: string
          nullable: true
...
    EpubNavigationSource:
      type: string
      enum: [epub3-nav, epub2-ncx, spine-fallback]
    EpubImportRole:
      type: string
      enum: [volume, part, chapter, section, frontmatter, backmatter, appendix, unknown]
...
    EpubContentRange:
      type: object
      required: [href]
      properties:
        href:
          type: string
        start_anchor:
          type: string
          nullable: true
        end_anchor:
          type: string
          nullable: true
    EpubNavigationNode:
      type: object
      required: [source_key, ordinal, depth, role, title, selected]
      properties:
        source_key:
          type: string
          description: Stable source SHA-256 plus normalized href/fragment identity.
        source_href:
          type: string
          nullable: true
        source_fragment:
          type: string
          nullable: true
        parent_source_key:
          type: string
          nullable: true
        ordinal:
          type: integer
          minimum: 0
        depth:
          type: integer
          minimum: 0
        role:
          $ref: '#/components/schemas/EpubImportRole'
        title:
          type: string
        linear:
          type: boolean
        selected:
          type: boolean
        content_ranges:
          type: array
          items:
            $ref: '#/components/schemas/EpubContentRange'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/EpubImportWarning'
        children:
          type: array
          items:
            $ref: '#/components/schemas/EpubNavigationNode'
    EpubCoverCandidate:
      type: object
      required: [source_path, media_type, size_bytes]
      properties:
        source_path:
          type: string
        media_type:
          type: string
        size_bytes:
          type: integer
          minimum: 0
        source_method:
          type: string
          enum: [epub3-cover-image, epub2-meta-cover, guide-cover, filename-fallback]
    EpubInspectionResponse:
      type: object
      required: [source_id, sha256, metadata, navigation_source, structure, warnings, limits]
      properties:
        source_id:
          type: string
          format: uuid
        sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
        duplicate_source:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
        cover:
          allOf:
            - $ref: '#/components/schemas/EpubCoverCandidate'
          nullable: true
        navigation_source:
          $ref: '#/components/schemas/EpubNavigationSource'
        structure:
          type: array
          items:
            $ref: '#/components/schemas/EpubNavigationNode'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/EpubImportWarning'
        limits:
          type: object
          additionalProperties:
            type: integer
...